requirements.txt
python-nmap
requests
PyYAML
vulners
beautifulsoup4
selenium
webdriver-manager
